+++ app/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod camera;
 pub mod star_system;
+mod utils;